safari-534.53-branch/LayoutTests/fast/dom/NodeList/nodelist-item-call-as-function-expected.txt

This tests that items in a NodeList can be retrieved directly by calling as a function with an integral index parameter, starting from 0.
It means NodeList[0] and NodeList(0) both work.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS nodeList[0] is nodeList(0)
PASS !nodeList[nodeList.length] is !nodeList(nodeList.length)
PASS successfullyParsed is true

TEST COMPLETE
